+++ fields/symmetries/dipole.cc
@@ -29,9 +29,9 @@ void asciiField::loadFieldMap_Dipole(gMappedField* map, double verbosity)
 
 	// Allocate memory. [LONGI][TRANSVERSE]
 	// as initialized in the map
-	map->B1_2D = new double*[map->np[0]];
+	map->B1_2D = new float*[map->np[0]];
 	for (unsigned int i = 0; i < map->np[0]; ++i)
-		map->B1_2D[i] = new double[map->np[1]];
+		map->B1_2D[i] = new float[map->np[1]];
 
 
 	double unit1 = get_number("1*" + map->getCoordinateWithSpeed(0).unit);
@@ -201,9 +201,3 @@ void gMappedField::GetFieldValue_Dipole( const double x[3], double *Bfield, int 
 
 	}
 }
-
-
-
-
-
-
